moved all packages into a PanCAD package so that the project can be built
