+++ src/PanCAD/svg/enum_color_keywords.py
