+++ src/PanCAD/svg/parsers.py
