+++ src/PanCAD/translators/freecad_sketcher_to_svg.py
@@ -7,8 +7,8 @@ this was dealing with FreeCAD and everything after this is dealing with
 SVG.
 """
 
-import trigonometry as trig
-import svg.generators as sg
+import PanCAD.trigonometry as trig
+import PanCAD.svg.generators as sg
 
 def line(freecad_properties: dict) -> dict:
     """Returns a dictionary of equivalent svg properties to recreate a
