+++ src/PanCAD/translators/svg_to_freecad_sketcher.py
@@ -7,8 +7,8 @@ pass' point, where all the work before this was dealing with SVG
 and everything after this is dealing with FreeCAD.
 """
 
-import trigonometry as trig
-import svg.parsers as sp
+import PanCAD.trigonometry as trig
+import PanCAD.svg.parsers as sp
 
 def line(svg_properties: dict) -> dict:
     """Returns a dictionary of equivalent FreeCAD properties to recreate
