+++ src/PanCAD/trigonometry.py
