+++ tests/test_config.py
@@ -5,7 +5,7 @@ import unittest
 
 sys.path.append('src')
 
-import config
+import PanCAD.config as config
 
 class TestConfig(unittest.TestCase):
     def setUp(self):
@@ -28,7 +28,6 @@ class TestConfig(unittest.TestCase):
     
     def test_read_defaults(self):
         settings = config.Config(self.NOMINAL_0)
-        print(settings.WINDOWS_SEARCH_PATHS)
     
     def test_read_invalid_options(self):
         with self.assertRaises(config.InvalidOptionError):
