+++ tests/test_svg_parsers.py
@@ -4,7 +4,7 @@ import unittest
 
 sys.path.append('src')
 
-import svg.parsers as sp
+import PanCAD.svg.parsers as sp
 
 class TestSVGPath(unittest.TestCase):
